requirements.txt
# Core Django
Django>=4.0,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
django-filter>=22.0
django-extensions>=3.2.0

# Database
psycopg2-binary>=2.9.0
django-tenants>=3.5.0

# Authentication & Security
djangorestframework-simplejwt>=5.0.0
django-allauth>=0.54.0
cryptography>=3.4.0

# Caching & Background Tasks
redis>=4.0.0
celery>=5.2.0
django-redis>=5.0.0
django-celery-beat>=2.4.0

# Payment Processing
stripe>=5.0.0

# API & Documentation
drf-spectacular>=0.25.0

# Monitoring & Logging
sentry-sdk>=1.30.0

# File Storage
boto3>=1.26.0
django-storages>=1.13.0

# Development
django-debug-toolbar>=4.0.0

# Production
gunicorn>=20.0.0
whitenoise>=6.0.0
django-environ>=0.10.0